refactor: DataFile 및 상속 엔티티 칼럼명, 디렉토리 위치 변경
- DataFile savedUrl => fileKey 이름 변경
- file/entity 디렉토리로 모든 엔티티 이동
