--- src/main/java/org/younginhambak/backend/file/DataFileStatus.java
@@ -1,5 +1,0 @@
-package org.younginhambak.backend.file;
-
-public enum DataFileStatus {
-  ACTIVE, DISABLED, DELETED, ARCHIVED
-}
