+++ src/main/java/org/younginhambak/backend/file/entity/DataFileStatus.java
@@ -1,0 +1,5 @@
+package org.younginhambak.backend.file.entity;
+
+public enum DataFileStatus {
+  ACTIVE, DISABLED, DELETED, ARCHIVED
+}
