+++ src/main/java/org/younginhambak/backend/file/entity/DocumentFile.java
@@ -1,4 +1,4 @@
-package org.younginhambak.backend.file;
+package org.younginhambak.backend.file.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -45,18 +45,18 @@ public class DocumentFile extends DataFile {
   /**
    * 새로운 DoucmentFile을 작성합니다.
    * @param fileName 파일 이름
-   * @param savedUrl 저장된 주소
+   * @param fileKey 저장된 주소
    * @param extension 피일 확장자
    * @return 새로 생성된 DocumentFile 객체
    */
   public static DocumentFile create(
           String fileName,
-          String savedUrl,
+          String fileKey,
           FileExtension extension
   ) {
     DocumentFile documentFile = new DocumentFile();
     documentFile.setFileName(fileName);
-    documentFile.setSavedUrl(savedUrl);
+    documentFile.setFileKey(fileKey);
     if (documentFile.isDocumentExtension(extension)) {
       documentFile.setExtension(extension);
     }
