+++ src/main/java/org/younginhambak/backend/file/entity/FileExtension.java
@@ -1,4 +1,4 @@
-package org.younginhambak.backend.file;
+package org.younginhambak.backend.file.entity;
 
 public enum FileExtension {
   HWP, DOC, PDF, TXT, PPT, MP3, MP4, JPG, PNG,
